+++ stories/Node.stories.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { NodeData } from '../src/types';
 import { Canvas } from '../src/Canvas';
 import { Node, Edge, MarkerArrow, Port, Icon, Arrow, Label, Remove, Add } from '../src/symbols';
-import { detectCircular, hasLink } from '../src/utils';
+import { detectCircular, hasLink } from '../src/helpers';
 
 export const LabelsAndIcons = () => (
   <div style={{ position: 'absolute', top: 0, bottom: 0, left: 0, right: 0 }}>